+++ decidim-core/app/packs/src/decidim/index.js
@@ -61,7 +61,7 @@ import addInputEmoji, { EmojiButton } from "src/decidim/input_emoji"
 import FocusGuard from "src/decidim/focus_guard"
 import backToListLink from "src/decidim/back_to_list"
 import markAsReadNotifications from "src/decidim/notifications"
-import RemoteModal from "src/decidim/ajax_modals"
+import RemoteModal from "src/decidim/remote_modal"
 import selectActiveIdentity from "src/decidim/identity_selector_dialog"
 import createTooltip from "src/decidim/tooltips"
 import createToggle from "src/decidim/toggle"
