+++ decidim-meetings/app/packs/stylesheets/decidim/meetings/_calendar.scss
