+++ decidim-meetings/app/packs/stylesheets/decidim/meetings/meetings.scss
@@ -1,4 +1,4 @@
 @import "stylesheets/decidim/meetings/list.scss";
 @import "stylesheets/decidim/meetings/item.scss";
-@import "stylesheets/decidim/meetings/year_calendar.scss";
+@import "stylesheets/decidim/meetings/calendar.scss";
 @import "stylesheets/decidim/meetings/live_event.scss";
